docs(filters): add documentation for filters and move the filter functions to a separate file
