+++ Flux.Net/Flux.Net/FluxFilter.cs
@@ -21,6 +21,11 @@ namespace Flux.Net
                 _stringBuilder.Append(" and ");
         }
 
+        /// <summary>
+        /// Add conditions to the filter predicate with raw Flux.
+        /// </summary>
+        /// <remarks>Records representing each row are passed as <c>r</c>.</remarks>
+        /// <param name="filters">A raw Flux predicate (eg. <c>"r._value &gt; 0 and r._value &lt; 50"</c>).</param>
         public FluxFilter Where(string filters)
         {
             AppendAndIfNeeded();
@@ -29,6 +34,10 @@ namespace Flux.Net
             return this;
         }
 
+        /// <summary>
+        /// Add a condition to the filter predicate to keep only records with the specified <paramref name="measurement"/>.
+        /// </summary>
+        /// <param name="measurement">Name of the measurement to filter records.</param>
         public FluxFilter Measurement(string measurement)
         {
             AppendAndIfNeeded();
@@ -37,6 +46,11 @@ namespace Flux.Net
             return this;
         }
 
+        /// <summary>
+        /// Add a condition to the filter predicate to keep only records with the specified <paramref name="tagKey"/> and <paramref name="tagValue"/>.
+        /// </summary>
+        /// <param name="tagKey">Key of the tag to filter.</param>
+        /// <param name="tagValue">Value of the tag to filter.</param>
         public FluxFilter Tag(string tagKey, string tagValue)
         {
             AppendAndIfNeeded();
@@ -45,6 +59,10 @@ namespace Flux.Net
             return this;
         }
 
+        /// <summary>
+        /// Add conditions to the filter predicate to keep only records with the specified <paramref name="tags"/>.
+        /// </summary>
+        /// <param name="tags">A dictionary of tag keys and values to filter records.</param>
         public FluxFilter Tags(IDictionary<string, string> tags)
         {
             if (tags == null)
@@ -56,6 +74,10 @@ namespace Flux.Net
             return this;
         }
 
+        /// <summary>
+        /// Add conditions to the filter predicate to keep only records with any specified <paramref name="fields"/>.
+        /// </summary>
+        /// <param name="fields">An array of field keys to filter records.</param>
         public FluxFilter Fields(params string[] fields)
         {
             if (fields == null || fields.Length == 0)
